fix: rename iconfield to labelfield, make field and labelfield exclusive
