+++ src/core/Input/shared/Field/styles.css
@@ -1,6 +1,6 @@
 /** @format */
 
-.d2ui-field-wrap {
+.d2ui-field {
     display: inline-flex;
     flex-direction: column;
     vertical-align: top;
