+++ src/core/Paper/index.js
@@ -8,23 +8,19 @@ import './styles.css'
 const bem = bemClassNames('paper')
 
 function Paper({ elevation, children, className }) {
-    const elevationClass = bem.b(`elevation-${elevation}`)
-    return (
-        <div className={`${elevationClass} ${className}`}>
-            {children}
-        </div>
-    )
+    const bemClass = bem.b(`elevation-${elevation}`)
+    return <div className={`${bemClass} ${className}`}>{children}</div>
 }
 
 Paper.propTypes = {
     elevation: PropTypes.oneOf([0, 1, 2, 3, 4, 6, 12, 24]),
-	className: PropTypes.string,
+    className: PropTypes.string,
     children: PropTypes.node,
 }
 
 Paper.defaultProps = {
     elevation: 1,
-	className: 'd2ui-paper-defaults',
+    className: 'd2ui-paper-defaults',
 }
 
 export { Paper }
